requirements.txt
os
numpy
pandas
textblob
textblob_fr
sklearn.metrics
sklearn.ensemble
sklearn.model_selection
sklearn.feature_extraction.text
sklearn.datasets
tweepy
nltk
nltk.stem.snowball
pickle
nltk.corpus
ltk.stem
re
tweepy.streaming
json
plotly.express
dash_bootstrap_components
dash.dependencies
dash
base64
multidict
wordcloud
PIL